chore: 전체 코드 리펙터링 및 기능추가

+++ src/main/kotlin/com/wooyassss/fileyassss/domain/file/dto/response/UploadFileResponse.kt
@@ -2,5 +2,5 @@ package com.wooyassss.fileyassss.domain.file.dto.response
 
 data class UploadFileResponse(
     val fileId: String,
-    val uploadProcess: Long
+    val uploadProcess: Double
 )
